Change temporarily to double preconditioning

+++ examples/test/driver_test.cu
@@ -33,7 +33,7 @@
 // Typedefs
 // -----------------------------------------------------------------------------
 typedef double REAL;
-typedef float  PREC_REAL;
+typedef double PREC_REAL;
 
 typedef typename cusp::csr_matrix<int, REAL, cusp::device_memory> Matrix;
 typedef typename cusp::array1d<REAL, cusp::device_memory>         Vector;
